refactor(cli): rename exp chats tui to exp agents
