+++ cli/root.go
@@ -164,7 +164,7 @@ func (r *RootCmd) AGPLExperimental() []*serpent.Command {
 		r.promptExample(),
 		r.rptyCommand(),
 		r.syncCommand(),
-		r.chatsCommand(),
+		r.agentsCommand(),
 	}
 }
